requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.4
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-filter==2.4.0
django-formtools==2.3
django-guardian==2.4.0
django-import-export==2.5.0
django-js-asset==1.2.2
django-reversion==3.0.9
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
et-xmlfile==1.1.0
future==0.18.2
httplib2==0.9.2
idna==2.10
importlib-metadata==4.5.0
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==2.0.1
mixins==0.1.4
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.2.0
PyJWT==1.7.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
serializers==0.2.4
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.1
